Feat: refactor packages to be compliant with lerna setup
