+++ packages/docs/package.json.lerna_backup
@@ -1,0 +1,154 @@
+{
+  "name": "react-ts-boilerplate-docs",
+  "version": "1.0.1",
+  "description": "Docs for the scalable-react-ts boilerplate",
+  "main": "index.js",
+  "author": "RyanCCollins",
+  "license": "MIT",
+  "engines": {
+    "node": "6.9.5",
+    "npm": "3.10.8"
+  },
+  "jest": {
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js"
+    ],
+    "moduleNameMapper": {
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/config/testing/__mocks__/fileMock.js",
+      "\\.(css|less)$": "<rootDir>/config/testing/__mocks__/styleMock.js"
+    },
+    "testPathIgnorePatterns": [
+      "<rootDir>/(build|docs|node_modules)/",
+      "/__mocks__/"
+    ],
+    "transform": {
+      ".*": "<rootDir>/config/testing/preprocessor.js"
+    },
+    "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$"
+  },
+  "scripts": {
+    "precommit": "npm run lint --silent && npm run test",
+    "build": "webpack --config ./webpack.config.js",
+    "start": "node devServer.js",
+    "predeploy": "rimraf build/public/*.*",
+    "deploy": "cross-env NODE_ENV=production ./node_modules/.bin/webpack --config ./webpack.config.prod.js",
+    "postdeploy": "npm run compile",
+    "compile": "./node_modules/.bin/tsc",
+    "setup:yarn": "yarn",
+    "preserve": "npm run compile",
+    "lint": "tslint \"packages/**/*.ts{,x}\" && tslint \"src/**/*.ts{,x}\"",
+    "serve": "cross-env NODE_ENV=server node server.js",
+    "postinstall": "npm run deploy",
+    "generator": "plop --plopfile ./config/generators/cli.js",
+    "test": "cross-env NODE_ENV=test jest",
+    "test:watch": "npm run test -- --watch",
+    "test:update": "npm run test -- --u",
+    "storybook": "start-storybook -p 9001 -c ./config/.storybook"
+  },
+  "dependencies": {
+    "@kadira/react-storybook-addon-info": "^3.3.0",
+    "@kadira/storybook": "^2.35.3",
+    "@kadira/storybook-addon-knobs": "^1.7.1",
+    "@types/body-parser": "^0.0.34",
+    "@types/compression": "latest",
+    "@types/express": "latest",
+    "@types/graphql": "latest",
+    "@types/isomorphic-fetch": "^0.0.33",
+    "@types/mongoose": "^4.7.7",
+    "@types/morgan": "latest",
+    "@types/node": "^6.0.64",
+    "@types/react": "^15.0.14",
+    "@types/react-dom": "^0.14.23",
+    "@types/react-ga": "^1.4.7",
+    "@types/react-redux": "^4.4.35",
+    "@types/react-router": "^3.0.6",
+    "@types/react-router-redux": "^4.0.42",
+    "@types/reselect": "^2.0.27",
+    "@types/sinon": "^1.16.35",
+    "@types/webpack": "^2.2.11",
+    "apollo-client": "^0.10.1",
+    "apollo-codegen": "^0.10.8",
+    "autoprefixer": "^6.7.6",
+    "axios": "^0.15.3",
+    "babel-polyfill": "^6.23.0",
+    "body-parser": "^1.17.0",
+    "compression": "^1.6.2",
+    "cors": "^2.8.1",
+    "exports-loader": "^0.6.4",
+    "express": "^4.15.0",
+    "github-markdown-css": "^2.4.1",
+    "graphql": "^0.9.1",
+    "graphql-server-express": "^0.6.0",
+    "graphql-tag": "^1.3.1",
+    "html-webpack-plugin": "^2.28.0",
+    "imports-loader": "^0.7.1",
+    "isomorphic-fetch": "^2.2.1",
+    "json-loader": "^0.5.4",
+    "markdown-loader": "^2.0.0",
+    "mongoose": "^4.8.5",
+    "morgan": "^1.8.1",
+    "node-env-file": "^0.1.8",
+    "openui": "1.0.1",
+    "plop": "^1.7.4",
+    "precss": "^1.4.0",
+    "react": "^15.4.1",
+    "react-addons-test-utils": "^15.4.2",
+    "react-apollo": "^0.13.2",
+    "react-dom": "^15.4.1",
+    "react-ga": "^2.1.2",
+    "react-hot-loader": "3.0.0-beta.6",
+    "react-markdown": "^2.4.5",
+    "react-redux": "^5.0.3",
+    "react-router": "^3.0.2",
+    "react-router-redux": "^4.0.8",
+    "redux": "^3.6.0",
+    "redux-connect": "^5.0.0",
+    "redux-logic": "^0.11.7",
+    "redux-mock-provider": "^1.0.0",
+    "reselect": "^2.5.4",
+    "rxjs": "^5.2.0",
+    "serialize-javascript": "^1.3.0",
+    "shortid": "^2.2.6",
+    "sinon": "^1.17.7",
+    "styled-components": "^1.4.4"
+  },
+  "devDependencies": {
+    "@types/enzyme": "^2.7.5",
+    "@types/jest": "^18.1.1",
+    "babel-core": "^6.23.1",
+    "babel-loader": "^6.3.2",
+    "babel-plugin-webpack-alias": "^2.1.2",
+    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-jest": "^19.0.0",
+    "babel-preset-react": "^6.23.0",
+    "babel-preset-stage-0": "^6.16.0",
+    "cross-env": "^3.2.3",
+    "css-loader": "^0.26.2",
+    "enzyme": "^2.7.1",
+    "enzyme-to-json": "^1.5.0",
+    "extract-text-webpack-plugin": "^2.0.0",
+    "html-loader": "^0.4.5",
+    "husky": "^0.13.2",
+    "jest": "^19.0.2",
+    "lerna": "2.0.0-beta.38",
+    "postcss-loader": "^1.3.3",
+    "react-storybook-addon-props-combinations": "^0.3.0",
+    "redux-logic-test": "^1.0.3",
+    "redux-mock-store": "^1.2.2",
+    "rimraf": "^2.6.1",
+    "style-loader": "^0.13.2",
+    "ts-jest": "^19.0.0",
+    "ts-loader": "^2.0.1",
+    "tslint": "^4.5.1",
+    "tslint-eslint-rules": "^3.5.1",
+    "tslint-loader": "^3.4.3",
+    "tslint-react": "^2.5.0",
+    "typescript": "^2.2.1",
+    "webpack": "^2.2.1",
+    "webpack-dev-middleware": "^1.10.1",
+    "webpack-hot-middleware": "^2.17.1",
+    "webpack-manifest-plugin": "^1.1.0"
+  }
+}
