+++ packages/ui/package.json
@@ -1,9 +1,9 @@
 {
   "name": "openui",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "description": "An open ui kit made with React and TypeScript",
   "main": "dist/index.js",
-  "typings": "dist/index.d.ts",
+  "types": "dist/index.d.ts",
   "repository": "https://github.com/scalable-react/scalable-react-typescript-boilerplate",
   "license": "MIT",
   "scripts": {
