+++ packages/ui/src/Markdown/index.tsx
@@ -9,6 +9,13 @@ export interface Props extends React.Props<typeof Markdown> {
 export default function Markdown({
   content,
 }: Props): JSX.Element {
+  if (typeof ReactMarkdown === 'undefined') {
+    return (
+      <Component className="markdown-body">
+        {content}
+      </Component>
+    );
+  }
   return (
     <Component className="markdown-body">
       <ReactMarkdown source={content} />
