+++ packages/ui/src/SvgIcon/types.ts
@@ -1,4 +1,4 @@
-export { Props } from './';
+export { Props } from './index';
 export interface SvgProps extends React.SVGProps {
   viewBox?: string;
 }
