+++ packages/ui/src/Toast/types.ts
@@ -1,3 +1,3 @@
-export { Props } from './';
+export { Props } from './index';
 
 export type Status = 'ok' | 'warning' | 'error' | 'none';
