+++ packages/ui/tsconfig.json
@@ -18,6 +18,7 @@
   "exclude": [
     "node_modules",
     "**/*.test.*",
-    "**/*.mock.*"
+    "**/*.mock.*",
+    "dist"
   ]
 }
